requirements/base.txt
pytz==2019.3
python-slugify==4.0.0
argon2-cffi==19.2.0
whitenoise==5.0.1
django-environ==0.4.5
django-model-utils==4.0.0
django-allauth==0.41.0
django-crispy-forms==1.8.1
djangorestframework==3.11.0
coreapi==2.3.3
Pillow==7.0.0
Django==3.0.3
ipdb==0.12.3
psycopg2-binary==2.8.4
factory-boy==2.12.0
django-extensions==2.2.6
Werkzeug==0.16.1
Sphinx==2.3.1
gunicorn==20.0.4
django-cors-middleware==1.5.0
django-cors-headers==3.2.1
django-rest-auth==0.9.5
djangorestframework-jwt==1.11.0
django-filter==2.2.0
regex==2020.4.4
mypy==0.761
django-stubs==1.4.0
pytest==5.3.5
pytest-sugar==0.9.2
flake8==3.7.9
black==19.10b0
pylint-django==2.0.13
pre-commit==2.0.1
pytest-django==3.8.0
django-debug-toolbar==2.2
django-coverage-plugin==1.8.0
django-countries==6.0
drf-yasg==1.17.1
six==1.14.0
django-oauth-toolkit==1.2.0
django-language-fields==1.0.0
numpy==1.18.2
sendgrid
django-money
razorpay
cryptography
django-taggit==1.3.0
boto3==1.14.60
django-storages==1.10
